requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
openai>=1.30
sentence-transformers>=2.7
feedparser>=6.0
httpx>=0.27
pydantic>=2.6
streamlit>=1.32
plotly>=5.20
pandas>=2.2
python-dotenv>=1.0
numpy>=1.26
reportlab>=4.0
